Completed Problem D561
Using DecimalFormat and its setRoundingMode method to handle 
display text. 
The type of number needs BigDecimal in Java8,because it is a bug 
in Java8 RoundingMode.HALF_UP.
For example 0.015 isbecoming 0.01, but expected 0.02.

What's different between "0.0" and "-0.0" ?
(http://www.programgo.com/article/4441958781/)
"-0.0" is produced when a floating-point operation results in a negative floating-point number so close to 0 that cannot be represented normally.

-2.0 / Float.POSITIVE_INFINITY  ->  -0.0

"-0.0" is numerically identical to "0.0". However, some operations involving "-0.0" are different than the same operation with "0.0".