Completed Problem B924
The fastest way is using DataInputStream and pre-buffer.
It is hard to find when has been reached the end of stream.
It is hard to implement, because handing data is very cumbersome.

The Second fast way is using BufferedReader and StringTokenizer or
split of String. But it consume too many memory.

The slowest way is using Scanner. It is too slow and consume too
many memory, although it is convenient to operate,liking next(),
nextInt(), nextLine(), and so on.